univer_django_docker_sqllite/requirements.txt
Django==2.1.1
djangorestframework
requests
lxml
bs4
